Renamed the text model attribute from text_transformers to text_model.
Removed PECore_Old as it isn't needed anymore.

Add a load_ckpt method to load the model weights after the model is instantiated, now the pretrained model can be either obtained by the from_config class methoc or by initializing a blank model and call the load_ckpt on it using the same args as the from_config method.