The build_cls_mask method can be removed as it compute the attention mask in case there is a cls token but in our specific case there isn't one. It is keep for the time being while considering if to delete it.
Added the prompt_forward method used for the SoftCPT implementation. The method takes as input directly the prompt embeddings and the tokenized input text and apply the forward logic to it. The tokenized input is used to get the EOS token that contains the pooled information as in our specific model there isn't an attention pooling mechanism as in the visual part of the PE.