Adds multi-dataset support
Adds a `MultiDataset` class to combine multiple datasets into one, allowing for easier handling of data from different sources.

Also, fixes an issue with image paths in the `BaseDataset` class by replacing backslashes with forward slashes for cross-platform compatibility.
Removes the identity attribute from the base dataset as it is not used.